backend/src/main/java/com/finance/project/dtos/dtos/AccountDTO.java
package com.finance.project.dtos.dtos;

import org.springframework.hateoas.RepresentationModel;

import java.util.Objects;

public class AccountDTO extends RepresentationModel<AccountDTO> {

    private String denomination;
    private String description;

    public AccountDTO(String denomination, String description) {
        this.denomination = denomination;
        this.description = description;
    }

    public String getDenomination() {
        return denomination;
    }

    public String getDescription() {
        return description;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        AccountDTO that = (AccountDTO) o;

        return Objects.equals(denomination, that.denomination) &&
                Objects.equals(description, that.description);
    }

    @Override
    public int hashCode() {
        return Objects.hash(denomination, description);
    }
}
